requirements.txt
discord.py
dateutil
